+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Add `channel_count` to `CommunitySummary` and `CommunitySummaryUpdates` ([#4129](https://github.com/open-chat-labs/open-chat/pull/4129))
+
 ## [[2.0.767](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.767-local_user_index)] - 2023-07-31
 
 ### Changed